Disable async to run functional tests
