+++ src/main/java/uk/gov/hmcts/cmc/claimstore/config/AsyncConfiguration.java
@@ -1,0 +1,24 @@
+package uk.gov.hmcts.cmc.claimstore.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.Executor;
+
+@Configuration
+@EnableAsync
+@Profile("!non-async")
+public class AsyncConfiguration {
+
+    @Bean(name = "threadPoolTaskExecutor")
+    public Executor threadPoolTaskExecutor(@Value("${async.max.threadPool.size}") Integer maxThreadPoolSize) {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setMaxPoolSize(maxThreadPoolSize);
+        executor.initialize();
+        return executor;
+    }
+}
