+++ src/main/java/uk/gov/hmcts/cmc/claimstore/config/NotificationsConfiguration.java
@@ -1,28 +1,14 @@
 package uk.gov.hmcts.cmc.claimstore.config;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.retry.annotation.EnableRetry;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import uk.gov.hmcts.cmc.claimstore.config.properties.notifications.NotificationsProperties;
 import uk.gov.service.notify.NotificationClient;
 
-import java.util.concurrent.Executor;
-
 @Configuration
 @EnableRetry
-@EnableAsync
 public class NotificationsConfiguration {
-
-    @Bean(name = "threadPoolTaskExecutor")
-    public Executor threadPoolTaskExecutor(@Value("${async.max.threadPool.size}") Integer maxThreadPoolSize) {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setMaxPoolSize(maxThreadPoolSize);
-        executor.initialize();
-        return executor;
-    }
 
     @Bean
     public NotificationClient notificationClient(NotificationsProperties notificationsProperties) {
